+++ src/seq_jacobi.hpp
@@ -5,6 +5,6 @@ using namespace std;
 #ifndef SEQ_JACOBI_H
 #define SEQ_JACOBI_H
 
-void seq_jacobi_method(vector<vector<double>> &A, vector<double> &b, double eps, int max_iterations, bool distance_stop, bool debug, vector<double> &res);
+void seq_jacobi_method(vector<vector<float>> &A, vector<float> &b, float eps, int max_iterations, bool distance_stop, bool debug, vector<float> &res);
 
 #endif
